Implement Get All Diary, fix Process, Add Weight, Update Weight Table,

+++ database/migrations/2022_06_06_065537_create_diaries_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('date');
             //$table->foreign('user_id')->references('id')->on('users');
             $table->foreign('process_id')->references('id')->on('processes');
-            $table->integer('weight_log')->nullable();
+            $table->float('weight_log')->nullable();
             $table->integer('heart_rate_log')->nullable();
             $table->integer('blood_pressure_log')->nullable();
             $table->tinyInteger('is_enough');
